Update and rename PUBG Discord Bot.py to PUBG_Discord.py
It's now a 'module' instead of built into a bot. 
Can do multiple players at one time. 
Should correct most errors and at least output something.
Outputs Errors given PUBGTracker. ie. The API is down, sometimes, or player cant be found.

Issues:
If your name is one of the game modes or regions, this won't work for you.
If the PUBGTracker API is down, it might not output anything.